requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.63.3
django-background-tasks==1.2.8
django-debug-toolbar==4.4.2
djangorestframework==3.15.2
fonttools==4.53.0
gunicorn==22.0.0
idna==3.7
isodate==0.6.1
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
python3-saml==1.16.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
xmlsec==1.3.14
